requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
dill==0.3.7
filelock==3.13.1
fsspec==2023.12.2
hydra-core==1.3.2
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
psutil==5.9.7
pynvml==11.5.0
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
pandas
transformers
